Don't ignore hidden files (see details)
Hidden files form part of the project, and they should be treated like
other regular files. Rather than ignoring hidden files, ignore just
those files (may be hidden or not) that really need to not be tracked.

Hidden files are as useful as other files: `.git`, `.gitignore`,
`.gitattributes`, `.npmignore`, to list a few.